ll40ls: added reset and backoff logic to driver
this adds automatic resets of the lidar when it becomes unresponsive,
and also tries to cope with NACKs from long acquisition times. It also
adds a "ll40ls regdump" command, and fixes "ll40ls reset" to reset the
sensor.

The default acquisition period is changed to 50ms, with backoff to
100ms on a NACK

Note that there are still times when the sensor can get into an
unrecoverable state.

Thanks to Austin, Dennis and Bob from PulseLight for assistance in
tracking down the problems.